Se agrega deletes

+++ app/src/main/java/com/rfz/appflotal/data/model/product/dto/ProductCrudDto.kt
@@ -16,5 +16,5 @@ data class ProductCrudDto (
     val loadCapacityId: Int,
 
     @SerializedName("fld_treadDepth")
-    val treadDepth: Float
+    val treadDepth: Int
 )
